emf_calculator/functions/Starter function/requirements.txt
appwrite
Pillow
tensorflow
opencv-python
gdown
mahotas
matplotlib

